+++ src/reinforce/wrappers/__init__.py
@@ -2,6 +2,7 @@
 
 from .async_vector import AsyncVectorEnv
 from .normalize import NormalizeObservation, NormalizeReward
+from .observation import FlattenObservation, FrameStack, OneHotObservation
 from .time_limit import TimeLimit
 from .vector import SyncVectorEnv
 
@@ -11,4 +12,7 @@ __all__ = [
     "NormalizeReward",
     "SyncVectorEnv",
     "AsyncVectorEnv",
+    "FrameStack",
+    "FlattenObservation",
+    "OneHotObservation",
 ]